+++ plugins/destination/clickhouse/typeconv/ch/definitions/definitions_test.go
@@ -1,4 +1,4 @@
-package ch
+package definitions
 
 import (
 	"strings"
